Merged file and memory store into atom store
This is still a work in progress. Combined/deduplicated the functionality
in the file-backed and memory-backed atom, and set a configuration flag we
can use to choose the store mechanism.

This still does not work correctly, because our schema requires the use of
Joda Time objects, which aren't compabible with enduro.  Hoping we can find
a better solution for time, such as JSR-310 which are now standard in Java
SE 8 and which replace Joda time.