Automatic datetime transformation for atom store
This update includes transformation for datetime objects in enduro
atoms, and hooks for performing transformations when enduro atoms
are read from and written to, extending the pattern we use for the
sql store.  Also includes updates to test infrastructure, but at
the moment the datetime objects continue to break many tests.

Continuing to remove memory-store code which is no longer needed.
This patch includes significant input from gbuisson and ssloan.