Integrated container monitoring via logspout
This update adds logspout functionality, which automatically routes
log output from all running containers (except logstash, which has
an exemption in the docker-compose.yml) to logstash, which is then
indexed by elasticsearch and made visible via kibana.  This allows
us to use kibana to monitor the state of the demo as well.